requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
xmltodict==0.13.0
cryptography==41.0.7
aiofiles==23.2.1
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
httpx==0.25.2
wechatpy==1.8.18
redis==5.0.1
celery==5.3.4 